Update requirements.txt
here is an updated version of the requirements that now can be installed. With the old version I got some errors on installing the package:
ERROR: Failed building wheel for ruamel.yaml
Failed to build ruamel.yaml
ERROR: Could not build wheels for ruamel.yaml, which is required to install pyproject.toml-based projects.

After updating the requirements to latest versions, it is working.